requirements.txt
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
azure-core>=1.30.0
python-dotenv>=1.0.0
tabulate>=0.9.0
Pillow>=10.0.0
